refactor(desktop): isolate workspace file service
